Replace generic FilterButton with TypeFilterButton to enablew setting of background with isActive
